+++ mmm_eval/configs/__init__.py
@@ -1,14 +1,9 @@
 from .config_registry import get_config
-from .configs import Config, EvalConfig, PyMCConfig
+from .configs import PyMCConfig
 from .rehydrators import PyMCConfigRehydrator
-from .utils import load_config, save_config
 
 __all__ = [
     "PyMCConfig",
-    "EvalConfig",
-    "Config",
     "get_config",
-    "save_config",
-    "load_config",
     "PyMCConfigRehydrator",
 ]